Fixing  Quantidade property type in ReceitaIngrediente Model

+++ GoodGood/Models/ReceitaIngrediente.cs
@@ -10,6 +10,6 @@ namespace GoodGood.Models {
         [NotNull]
         public  Ingrediente? Ingrediente {get; set;}
 
-        public int Quantidade {get; set;}
+        public float Quantidade {get; set;}
     }
 }
